+++ src/main/java/kitchenpos/eatinorder/infra/JpaEatInMenuRepository.java
@@ -10,7 +10,4 @@ import java.util.List;
 import java.util.UUID;
 
 public interface JpaEatInMenuRepository extends EatInMenuRepository, JpaRepository<EatInMenu, UUID> {
-    @Query("select m from EatInMenu m join m.menuProducts mp where mp.product.id = :productId")
-    @Override
-    List<EatInMenu> findAllByProductId(@Param("productId") UUID productId);
 }
