+++ src/main/java/kitchenpos/takeoutorder/domain/TakeOutMenuRepository.java
@@ -13,6 +13,5 @@ public interface TakeOutMenuRepository {
 
     List<TakeOutMenu> findAllByIdIn(List<UUID> ids);
 
-    List<TakeOutMenu> findAllByProductId(UUID productId);
 }
