+++ src/test/java/kitchenpos/application/takeout/Fixtures.java
@@ -1,0 +1,56 @@
+package kitchenpos.application.takeout;
+
+import kitchenpos.common.OrderStatus;
+import kitchenpos.common.OrderType;
+import kitchenpos.eatinorder.domain.EatInMenu;
+import kitchenpos.eatinorder.domain.EatInOrder;
+import kitchenpos.eatinorder.domain.EatInOrderLineItem;
+import kitchenpos.eatinorder.domain.OrderTable;
+import kitchenpos.takeoutorder.domain.TakeOutMenu;
+import kitchenpos.takeoutorder.domain.TakeOutOrder;
+import kitchenpos.takeoutorder.domain.TakeOutOrderLineItem;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.Random;
+import java.util.UUID;
+
+public class Fixtures {
+    public static final UUID INVALID_ID = new UUID(0L, 0L);
+
+    public static OrderTable orderTable(final boolean occupied, final int numberOfGuests) {
+        final OrderTable orderTable = new OrderTable();
+        orderTable.setId(UUID.randomUUID());
+        orderTable.setName("1번");
+        orderTable.setNumberOfGuests(numberOfGuests);
+        orderTable.setOccupied(occupied);
+        return orderTable;
+    }
+
+    public static TakeOutOrder order(final OrderStatus status) {
+        final TakeOutOrder order = new TakeOutOrder();
+        order.setId(UUID.randomUUID());
+        order.setType(OrderType.EAT_IN);
+        order.setStatus(status);
+        order.setOrderDateTime(LocalDateTime.of(2020, 1, 1, 12, 0));
+        order.setOrderLineItems(Arrays.asList(orderLineItem()));
+        return order;
+    }
+
+    public static TakeOutOrderLineItem orderLineItem() {
+        final TakeOutOrderLineItem orderLineItem = new TakeOutOrderLineItem();
+        orderLineItem.setSeq(new Random().nextLong());
+        orderLineItem.setMenu(menu(20_000L, true));
+        return orderLineItem;
+    }
+
+    public static TakeOutMenu menu(final long price, final boolean displayed) {
+        final TakeOutMenu menu = new TakeOutMenu();
+        menu.setId(UUID.randomUUID());
+        menu.setName("후라이드+후라이드");
+        menu.setPrice(BigDecimal.valueOf(price));
+        menu.setDisplayed(displayed);
+        return menu;
+    }
+}
